third_party/blink/web_tests/fast/css/css-namespace-rule-expected.txt

Check the basic attribute support of CSSOM CSSNamespaceRule

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS CSSRule.NAMESPACE_RULE is 10
PASS namespacerule.type is 10
PASS namespacerule.namespaceURI is "http://www.w3.org/1999/xhtml"
PASS namespacerule.prefix is "my-namespace"
PASS defaultNamespacerule.cssText is "@namespace url(\"default\");"
PASS namespacerule.cssText is "@namespace my-namespace url(\"http://www.w3.org/1999/xhtml\");"
PASS getComputedStyle(document.body).color is "rgb(255, 0, 0)"
Deleting namespace rule when other rules are present should throw InvalidStateException.
WARN: shouldBe() expects string arguments
PASS 11 is 11
Inserting new namespace rule when other rules are present should throw InvalidStateException.
WARN: shouldBe() expects string arguments
PASS 11 is 11
PASS successfullyParsed is true

TEST COMPLETE

Here is some text!
